+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 4.1.1
+* add key related model funcs
+* fix bug with 'additionalProperties'. From JSON schema specs this value can be boolean or object. The boolean version can't used before.
+
 # 4.1.0
 * add enum values